Hard limit on total number of operations
Rather than pass a current number of operations around which was hard to
track, or check the total filter ops at the end, which might mean wasted
processing, we check the number of child ops at different points of processing.

Added a new op_count() method to FilterOp to aid with that.