Only index fields defined in the search schema
This is not the current behaviour where all extras and custom fields are
indexed by default and it might take a bit of adjustement. But perhaps
we can make it easier at the scheming level automatically adding custom
fields to the search schema unless explictly turned off.

TODO: include fields from ISearchFeature plugins. This might be entity
specific