Add validator to allow dict fields in a schema
Navl has historically not supported fields that have dicts as value.
Defining a subschema and trying to validate a dict directly results in
the dreaded exception:

    DataError: Only lists of dicts can be placed against subschema

So you need to always pass a list of dicts:

    {
        "facets": [
            {"field": "tags"}
        ]
    }

Instead of a dict, which feels really akward and complicates
implementations:

    {
        "facets": {
            "field": "tags"
        }
    }

The `dict_subschema()` validator (name provisional) allows to define a
second schema against which the data contained in the field (a dict with
field keys) needs to be validated. If errors are found, these are added
to the parent field key, with the sub field included in the error
message. If there are no errors, the validated sub data is assigned to the
parent field.